requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
requests
pandas
pydantic
pypdf
langchain-groq
langchain-core
